AgentOps/agents/requirements.txt
# AgentOps — Python Requirements
# All agentic AI frameworks and dependencies

# ─── Core Agent Frameworks ───────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.20.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
pyautogen>=0.4.0                 # Microsoft AutoGen
agentscope>=0.1.0                # Alibaba AgentScope

# ─── LLM Providers ───────────────────────────────────────────────────────────
openai>=1.50.0
anthropic>=0.30.0
langchain-anthropic>=0.3.0
ollama>=0.3.0

# ─── Vector Stores & RAG ─────────────────────────────────────────────────────
qdrant-client>=1.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0

# ─── Observability & Tracing ─────────────────────────────────────────────────
langsmith>=0.1.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
prometheus-client>=0.21.0

# ─── Data & Storage ──────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
minio>=7.2.0

# ─── Workflow Orchestration ──────────────────────────────────────────────────
temporalio>=1.7.0
celery>=5.4.0

# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0

# ─── Security ────────────────────────────────────────────────────────────────
hvac>=2.3.0                      # HashiCorp Vault
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0

# ─── Utilities ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
structlog>=24.4.0
tenacity>=9.0.0
psutil>=6.0.0
